backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
openai>=1.0.0
psycopg2-binary
python-dotenv
